% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{arrange_ggplot2}
\alias{arrange_ggplot2}
\title{Arrange ggplots}
\usage{
arrange_ggplot2(..., nrow = NULL, ncol = NULL, as.table = FALSE)
}
\arguments{
\item{...}{The ggplots.}

\item{nrow}{Number of row the ggplots are wanted to be arranged.}

\item{ncol}{Number of column the ggplots are wanted to be arranged.}

\item{as.table}{Logical.}
}
\value{
Plots.
}
\description{
Function for arranging ggplots. use png(); arrange(p1, p2, ncol=1); dev.off() to save.
}

