% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{check.cor}
\alias{check.cor}
\title{Check the correlations between the variables}
\usage{
check.cor(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
Print out top 20 correlation rate and export correlation rate of all variables to *.csv file.
}
\description{
Check the correlations between the variables
}

