% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{co}
\alias{co}
\title{Carryover transformation}
\usage{
co(x, co.rate)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{co.rate}{The carry over rate. The value should be between 0 and 1.}
}
\value{
The transformed numeric vector.
}
\description{
Formula: period2 = period1 * carryover rate + period2
}

