% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{cp}
\alias{cp}
\title{Carryover + Power-curve transformation}
\usage{
cp(x, co.rate, exponent)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{co.rate}{The carry over rate. The value should be between 0 and 1.}

\item{exponent}{The exponent rate.}
}
\value{
The transformed numeric vector.
}
\description{
x <- pc(co(x, co.rate), exponent)
}
\seealso{
\code{\link{co}}
}

