% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{cs}
\alias{cs}
\title{Carryover + S-curve}
\usage{
cs(x, co.rate, lamda1, lamda2)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{co.rate}{The carry over rate. The value should be between 0 and 1.}

\item{lamda1, lamda2}{Two parameters of the s-curve.}
}
\value{
The transformed numeric vector.
}
\description{
Formula: transformed_x <- sc(co(x, co.rate), lamda1, lamda2)
}
\seealso{
\code{\link{co}} and \code{\link{sc}}
}

