% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{final.output}
\alias{final.output}
\title{Final output}
\usage{
final.output(resp, data, rawdata, tvar, fit, prmt, aic = FALSE)
}
\arguments{
\item{resp}{A character, name of response variable.}

\item{data}{The data.frame of the modified data.}

\item{tvar}{A character, name of time variable, used for plotting.}

\item{fit}{The temporary model.}

\item{prmt}{A dataframe including names of predictors, transforming parameters and variable status("alive" or "dead").}

\item{aic}{Logical value. TRUE for need AIC step session, FALSE for not. The default value is FALSE.}
}
\value{
Print output to the screen. External files such as *.csv files or *.xlsx files.
}
\description{
Print summary and some useful statistics of the temporary model to the screen.
Output necessary data to external files such as *.csv files for rebuilding models.
}

