% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{hmodif}
\alias{hmodif}
\title{Holiday variable transformation}
\usage{
hmodif(pred, resp, data)
}
\arguments{
\item{pred}{A character, name of variable to be transformed.}

\item{resp}{A character, name of the response variable of the model.}

\item{data}{A data frame to store the variable transformed (original variable data will be replaced).}
}
\value{
The complete data.frame with selected variable transformed.
}
\description{
Holiday vector equals to 1 on the current period of the holiday and 0 on other periods. The variable is replaced by the transformed data according to the current scale of the response variable.
}

