% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{loop.output}
\alias{loop.output}
\title{Loop output}
\usage{
loop.output(resp, data, fit, pred, tvar)
}
\arguments{
\item{resp}{A character, name of response variable.}

\item{data}{The data.frame of the modified data.}

\item{fit}{The temporary model.}

\item{pred}{A character, name of predictor.}

\item{tvar}{A character, name of time variable, used for plotting.}
}
\value{
Print output to the screen.
}
\description{
Print summary and some useful statistics of the temporary model to the screen.
}

