\name{mmm-package}
\alias{mmm-package}
\alias{mmm-package}
\docType{package}
\title{
mmm: an R package for analyzing multivariate longitudinal data with multivariate marginal models
}
\description{
fits multivariate marginal models for multivariate longitudinal data for both continuous and discrete responses
}
\details{
\tabular{ll}{
Package: \tab mmm\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2013-18-06\cr
License: \tab GPL (>=2)\cr
}
}