% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{modif}
\alias{modif}
\title{Variable transformation}
\usage{
modif(pred, resp, data, tm.prmt)
}
\arguments{
\item{pred}{A character, name of variable to be transformed.}

\item{resp}{A character, name of the response variable of the model.}

\item{data}{A data frame to store the variable transformed (original variable data will be replaced).}

\item{tm.prmt}{A numeric vector of length 4 containing transformation parameters.}
}
\value{
The complete data.frame with selected variable transformed.
}
\description{
The variable is replaced by the transformed data using carry over and S-curve or power-curve transformation.
}

