% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{rebuild}
\alias{rebuild}
\title{Rebuild model}
\usage{
rebuild(resp, data, st.row)
}
\arguments{
\item{resp}{A character, name of response variable.}

\item{data}{Raw data.frame without modification.}

\item{st.row}{The number of modeling start row.}
}
\value{
The transformed data and base model to the package working environment.
}
\description{
After importing the file of the variables and correspondent transforming parameters, rebuild the model of the specified response variable.
}

