% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{sc}
\alias{sc}
\title{S-curve transformation}
\usage{
sc(x, lamda1, lamda2)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{lamda1, lamda2}{Two parameters of the s-curve.}
}
\value{
The transformed numeric vector.
}
\description{
Formula: 1-e^(-lamda1 * x^lamda2)
}

