% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{StepReg}
\alias{StepReg}
\title{StepReg}
\usage{
StepReg(...)
}
\arguments{
\item{...}{Nothing necessary.}
}
\value{
Model and related files.
}
\description{
The structure function which cover the whole functionality using all self-built functions.
}

