% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{trial}
\alias{trial}
\title{Linear Regression Modeling}
\usage{
trial(data, resp, fit = NULL, action = 1, pred)
}
\arguments{
\item{data}{A data.frame contain both predictor and respondent variables.}

\item{resp}{A character, name of respondent variable included in the data.frame.}

\item{fit}{A linear regression model. The default value is NULL.}

\item{action}{1 or -1. 1 means add a predictor; -1 means delete a predictor. The default value is 1.}

\item{pred}{A character, name of predictor variable included in the data.frame.}
}
\value{
The new updated model.
}
\description{
Update a linear regression model by adding or removing one new variable.
Create a model if there is not one at all.
}

