% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\name{warn}
\alias{warn}
\title{Warning users}
\usage{
warn(fit1, fit2, p.cons = 0.2)
}
\arguments{
\item{fit1, fit2}{Two models that are needed to be compared.}

\item{p.cons}{The critical point, when higher than which warnings appear.}
}
\value{
Warnings printed on the screen.
}
\description{
1. Warn user when there is some coefficient which changes **sign** between two models;
2. Warn user when there is big gap between **p-value** of one pred in two models.
}

