\name{mmm-package}
\alias{mmm-package}
\alias{mmm-package}
\docType{package}
\title{
Multivariate Marginal Models with Response Specific Parameters
}
\description{
fits multivariate marginal models with response specific parameters for multivariate longitudinal data for both continuous and discrete responses
}
\details{
\tabular{ll}{
Package: \tab mmm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-22\cr
License: \tab GPL (>=2)\cr
}
}