\name{mmm-package}
\alias{mmm-package}
\alias{mmm-package}
\docType{package}
\title{
mmm: an R package for analyzing multivariate longitudinal data with multivariate marginal models
}
\description{
fits multivariate marginal models for multivariate longitudinal data for both continuous and discrete responses

Asar, O., Ilk, O. (2013). mmm: an R package for analyzing multivariate longitudinal data with multivariate marginal models, 
\emph{Computer Methods and Programs in Biomedicine}, \bold{112}, 649-654.

}
\details{
\tabular{ll}{
Package: \tab mmm\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2013-01-01\cr
License: \tab GPL (>=2)\cr
}
}