\name{mmm-package}
\alias{mmm-package}
\alias{mmm-package}
\docType{package}
\title{
Multivariate Marginal Models
}
\description{
fits multivariate marginal models for multivariate longitudinal data for both continuous and discrete responses
}
\details{
\tabular{ll}{
Package: \tab mmm\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2011-09-26\cr
License: \tab GPL (>=2)\cr
}
}